+++ Formula/sec-agent.rb
@@ -1,9 +1,9 @@
 class SecAgent < Formula
   desc "macOS Enclave-Bound Session Agent for Encrypted Secrets"
   homepage "https://github.com/iafilius/sec-agent"
-  url "https://github.com/iafilius/sec-agent/releases/download/v2.13.3/sec-agent_v2.13.3_darwin_arm64.tar.gz"
-  version "2.13.3"
-  sha256 "275560f5314d8dfc9668a5c13adca6d99b383e1aff635a6815c74980b7a118b4"
+  url "https://github.com/iafilius/sec-agent/releases/download/v2.13.4/sec-agent_v2.13.4_darwin_arm64.tar.gz"
+  version "2.13.4"
+  sha256 "16261fd17eb127977639cca44acd05a20d054d1d6c021e1c860f7572b8a0a79b"
   license "GPL-3.0-or-later"
 
   depends_on :macos